+++ src/app/view/admin/admin-alldevices/admin-alldevices.component.html
@@ -22,7 +22,7 @@
     <hr class="mt-0 pt-0" />
 
     <div *ngIf="!hideMap">
-      <app-simple-map></app-simple-map>
+      <app-map-devices></app-map-devices>
     </div>
 
     <div class="d-flex justify-content-end gap-2 my-4 mx-4">
