+++ src/app/view/map/map-devices.component.html
@@ -1,0 +1,8 @@
+<div class="map-container">
+  <div
+    class="map"
+    leaflet
+    [leafletOptions]="options"
+    (leafletMapReady)="onMapReady($event)"
+  ></div>
+</div>
