+++ src/app/view/map/map-devices.component.scss
@@ -1,0 +1,12 @@
+.map-container {
+  height: 400px;
+  width: 100%;
+  border-radius: 8px;
+  overflow: hidden;
+  box-shadow: 0 2px 5px rgba(0, 0, 0, 0.15);
+}
+
+.map {
+  height: 100%;
+  width: 100%;
+}
