+++ src/app/view/map/models/device.model.ts
@@ -1,0 +1,6 @@
+export interface Device {
+  id: string;
+  externalId: string;
+  latitude: number;
+  longitude: number;
+}
